BLGMIGRATION-866 remove: result doesn't have benefits to define
