+++ core/src/main/scala/com/nulabinc/backlog/migration/common/dsl/AppDSL.scala
@@ -1,10 +1,9 @@
 package com.nulabinc.backlog.migration.common.dsl
 
-import com.nulabinc.backlog.migration.common.shared.Result.Result
 import simulacrum.typeclass
 
 @typeclass
 trait AppDSL[F[_]] {
   def pure[A](a: A): F[A]
-  def fromError[E, A](error: E): F[Result[E, A]]
+  def fromError[E, A](error: E): F[Either[E, A]]
 }
