+++ core/src/main/scala/com/nulabinc/backlog/migration/common/interpreters/TaskAppDSL.scala
@@ -1,7 +1,6 @@
 package com.nulabinc.backlog.migration.common.interpreters
 
 import com.nulabinc.backlog.migration.common.dsl.AppDSL
-import com.nulabinc.backlog.migration.common.shared.Result.Result
 import monix.eval.Task
 
 case class TaskAppDSL() extends AppDSL[Task] {
@@ -9,6 +8,6 @@ case class TaskAppDSL() extends AppDSL[Task] {
   override def pure[A](a: A): Task[A] =
     Task(a)
 
-  override def fromError[E, A](error: E): Task[Result[E, A]] =
+  override def fromError[E, A](error: E): Task[Either[E, A]] =
     Task(Left(error))
 }
