+++ core/src/main/scala/com/nulabinc/backlog/migration/common/shared/Result.scala
@@ -5,35 +5,35 @@ import monix.eval.Task
 
 
 object Result {
-  type Result[E, A] = Either[E, A]
 
-  def success[A](a: A): Result[Nothing, A] = Right(a)
-  def error[E, A](error: E): Result[E, A] = Left(error)
-}
+  type Result[E, A] = Task[Either[E, A]]
 
-object AsyncResult {
-  import Result.Result
+  def success[A](a: A): Result[Nothing, A] = Task(Right(a))
 
-  type AsyncResult[E, A] = Task[Result[E, A]]
+  def error[E, A](error: E): Result[E, A] = Task(Left(error))
 
-  def success[A](a: A): AsyncResult[Nothing, A] = Task(Result.success(a))
-
-  def error[E, A](error: E): AsyncResult[E, A] = Task(Result.error(error))
-
-  def fromEither[E, A](result: Either[E, A]): AsyncResult[E, A] = Task(result)
+  def fromEither[E, A](result: Either[E, A]): Result[E, A] = Task(result)
 
   object syntax {
 
-    implicit class AsyncResultOps[E, A](result: AsyncResult[E, A]) {
+    implicit class AsyncResultOps[E, A](result: Result[E, A]) {
       def handleError: EitherT[Task, E, A] =
         EitherT(result)
 
-      def mapError[E2](f: E => E2): AsyncResult[E2, A] =
+      def mapError[E2](f: E => E2): Result[E2, A] =
         result.flatMap { inner =>
           inner.fold(
-            error => AsyncResult.error(f(error)),
-            data => AsyncResult.success(data)
+            error => Result.error(f(error)),
+            data => Result.success(data)
           )
+        }
+    }
+
+    implicit class OptionTaskOps[A](optTaskValue: Option[Task[A]]) {
+      def sequence: Task[Option[A]] =
+        optTaskValue match {
+          case Some(task) => task.map(Some(_))
+          case None => Task(None)
         }
     }
