+++ core/src/main/scala/com/nulabinc/backlog/migration/common/shared/syntax.scala
@@ -2,12 +2,10 @@ package com.nulabinc.backlog.migration.common.shared
 
 import cats.Monad
 import cats.data.EitherT
-import com.nulabinc.backlog.migration.common.shared.Result.Result
-import monix.eval.Task
 
 object syntax {
 
-  implicit class AsyncResultOps[F[_]: Monad, E, A](result: F[Result[E, A]]) {
+  implicit class ResultOps[F[_]: Monad, E, A](result: F[Either[E, A]]) {
     def handleError: EitherT[F, E, A] =
       EitherT(result)
 
@@ -20,11 +18,5 @@ object syntax {
 //      }
   }
 
-  implicit class OptionTaskOps[A](optTaskValue: Option[Task[A]]) {
-    def sequence: Task[Option[A]] =
-      optTaskValue match {
-        case Some(task) => task.map(Some(_))
-        case None => Task(None)
-      }
-  }
+
 }
